add logic for plus minus buttons
