+++ app/src/main/java/com/example/deliveryfood/data/ResponseDishes.kt
@@ -1,7 +1,7 @@
 package com.example.deliveryfood.data
 
-import com.google.gson.annotations.SerializedName
 import com.google.gson.annotations.Expose
+import com.google.gson.annotations.SerializedName
 
 data class ResponseDishes(
     @SerializedName("dishes")
