+++ app/src/main/java/com/example/deliveryfood/fragment/AccountFragment.kt
@@ -1,0 +1,17 @@
+package com.example.deliveryfood.fragment
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.deliveryfood.R
+
+class AccountFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_account, container, false)
+    }
+}
