+++ app/src/main/java/com/example/deliveryfood/fragment/SearchFragment.kt
@@ -1,0 +1,17 @@
+package com.example.deliveryfood.fragment
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.deliveryfood.R
+
+class SearchFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_search, container, false)
+    }
+}
