+++ app/src/main/java/com/example/deliveryfood/fragment_popup/PopUpOrderFragment.kt
@@ -1,20 +1,16 @@
-package com.example.deliveryfood
+package com.example.deliveryfood.fragment_popup
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.activityViewModels
-import com.example.deliveryfood.adapter.OrderAdapter
 import com.example.deliveryfood.data.OrderItem
-import com.example.deliveryfood.databinding.FragmentMainBinding
 import com.example.deliveryfood.databinding.FragmentPopUpOrderBinding
 import com.example.deliveryfood.view.DishViewModel
 import com.example.deliveryfood.view.OrderViewModel
 import com.squareup.picasso.Picasso
-import io.reactivex.rxjava3.internal.subscriptions.SubscriptionHelper.cancel
 
 
 class PopUpOrderFragment : DialogFragment() {
