+++ app/src/main/res/layout/fragment_category.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingLeft="16dp"
-    tools:context=".CategoryFragment">
+    tools:context=".fragment.CategoryFragment">
 
     <FrameLayout
         android:id="@+id/toolbar"
@@ -36,7 +36,6 @@
         android:layout_marginTop="15dp"
         android:orientation="horizontal"
         app:layout_constraintEnd_toStartOf="@+id/guideline_right"
-        app:layout_constraintHorizontal_bias="0.513"
         app:layout_constraintStart_toStartOf="@+id/guideline_left"
         app:layout_constraintTop_toBottomOf="@+id/toolbar" />
 
@@ -48,6 +47,5 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toStartOf="@+id/guideline_right"
         app:layout_constraintStart_toStartOf="@+id/guideline_left"
-        app:layout_constraintTop_toBottomOf="@+id/recycler_tag"
-        app:layout_constraintVertical_bias="0.0" />
+        app:layout_constraintTop_toBottomOf="@+id/recycler_tag" />
 </androidx.constraintlayout.widget.ConstraintLayout>
