+++ app/src/main/res/layout/fragment_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainFragment">
+    tools:context=".fragment.MainFragment">
 
     <FrameLayout
         android:id="@+id/actionbar"
@@ -25,5 +25,4 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/actionbar" />
-
 </androidx.constraintlayout.widget.ConstraintLayout>
