+++ app/src/main/res/layout/item_cuisine_layout.xml
@@ -27,11 +27,10 @@
         android:layout_marginTop="12dp"
         android:layout_marginBottom="86dp"
         android:minLines="2"
-        android:text="Пекарни и кондитерские"
+        android:text="@string/title_cuisine"
         android:textSize="20sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintStart_toStartOf="@+id/image_cuisine"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.571" />
+        app:layout_constraintTop_toTopOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
