+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/mainFragment"
-        android:name="com.example.deliveryfood.MainFragment"
+        android:name="com.example.deliveryfood.fragment.MainFragment"
         android:label="fragment_main"
         tools:layout="@layout/fragment_main" >
         <action
@@ -16,22 +16,22 @@
     </fragment>
     <fragment
         android:id="@+id/accountFragment"
-        android:name="com.example.deliveryfood.AccountFragment"
+        android:name="com.example.deliveryfood.fragment.AccountFragment"
         android:label="fragment_account"
         tools:layout="@layout/fragment_account" />
     <fragment
         android:id="@+id/basketFragment"
-        android:name="com.example.deliveryfood.BasketFragment"
+        android:name="com.example.deliveryfood.fragment.BasketFragment"
         android:label="fragment_basket"
         tools:layout="@layout/fragment_basket" />
     <fragment
         android:id="@+id/searchFragment"
-        android:name="com.example.deliveryfood.SearchFragment"
+        android:name="com.example.deliveryfood.fragment.SearchFragment"
         android:label="fragment_search"
         tools:layout="@layout/fragment_search" />
     <fragment
         android:id="@+id/categoryFragment"
-        android:name="com.example.deliveryfood.CategoryFragment"
+        android:name="com.example.deliveryfood.fragment.CategoryFragment"
         android:label="fragment_category"
         tools:layout="@layout/fragment_category" >
         <action
